requirements.txt
numpy>=1.24.0
flask>=2.0.0
folium>=0.12.0
geopy>=2.2.0
openrouteservice>=2.3.3
python-dotenv>=0.19.0
requests>=2.26.0
